Update AWS sdk to 2.6.15

+++ apps/files_external/3rdparty/aws-sdk-php/Guzzle/Service/Description/SchemaFormatter.php
@@ -109,7 +109,7 @@ class SchemaFormatter
      */
     public static function formatTimestamp($value)
     {
-        return self::dateFormatter($value, 'U');
+        return (int) self::dateFormatter($value, 'U');
     }
 
     /**
